refactor: extract age-filter auto-actions from content.ts to lib/auto-actions.ts (Phase 4 Batch 1)
Extract 7 age-filter orchestration functions (setAutoAgeAttributes,
findAgeFilterBlock, openAgeFilterDropdown, resolveAgeSelectBox,
waitForAgeFilterDropdown, resolveAgeFilterActions, autoApplyAgeFilterFromUrl)
from content.ts into auto-actions.ts using the established DI factory pattern.

createAutoActions(deps) accepts 16 injected dependencies and returns 6
exported functions. The internal setAutoAgeAttributes is not returned.
Shared UI helpers (activateElement, fireMouseEvent, setInputValue, etc.)
remain in content.ts and are passed as deps.

Part of ~1,215L reduction target for content.ts.
